refactor: 클래스 위치 재배치
