--- src/main/java/domain/strategy/DeckSettingStrategy.java
@@ -1,11 +1,0 @@
-package domain.strategy;
-
-import domain.Card;
-import domain.Cards;
-
-import java.util.List;
-
-public interface DeckSettingStrategy {
-
-    Cards initialize();
-}
