+++ src/main/java/participant/Dealer.java
@@ -1,4 +1,9 @@
-package domain;
+package participant;
+
+import card.Card;
+import card.Cards;
+import card.Deck;
+import strategy.DeckShuffleStrategy;
 
 import java.util.Comparator;
 import java.util.List;
@@ -9,10 +14,12 @@ public class Dealer {
     private static final int INITIAL_CARD_COUNT = 2;
 
     private final Cards cards;
+    private final Deck deck;
 
     public Dealer(Cards cards) {
         validateInitialCardsSize(cards);
         this.cards = cards;
+        this.deck = new Deck(new DeckShuffleStrategy());
     }
 
     public Card openOneCard() {
@@ -35,6 +42,10 @@ public class Dealer {
 
     public int sumCardNumbers() {
         return cards.sumCardNumbers();
+    }
+
+    public boolean isBlackJack() {
+        return cards.isBlackJack();
     }
 
     private void validateInitialCardsSize(Cards cards) {
