+++ src/main/java/strategy/DeckSettingStrategy.java
@@ -1,0 +1,8 @@
+package strategy;
+
+import card.Cards;
+
+public interface DeckSettingStrategy {
+
+    Cards initialize();
+}
