+++ src/test/java/card/CardsTest.java
@@ -1,9 +1,9 @@
-package domain;
+package card;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import domain.constant.TrumpEmblem;
-import domain.constant.TrumpNumber;
+import constant.TrumpEmblem;
+import constant.TrumpNumber;
 import java.util.ArrayList;
 import java.util.List;
 import org.junit.jupiter.params.ParameterizedTest;
